requirements.txt
beautifulsoup4==4.13.5
Django==5.2
djangorestframework==3.16.0
feedparser==6.0.11
python-decouple==3.8
Requests==2.32.5
transformers==4.55.4
django-crontab
psycopg
psycopg-binary
transformers
torch
keras==2.11.0
Pillow